fix the issue of frequent poi update in iPhone resolve issue #6
