+++ lAyeR/lAyeR/ARViewController_CameraBackgroundExtension.swift
@@ -1,0 +1,73 @@
+//
+//  ARViewController_CameraBackgroundExtension.swift
+//  lAyeR
+//
+//  Created by luoyuyang on 22/03/17.
+//  Copyright © 2017年 nus.cs3217.layer. All rights reserved.
+//
+
+import AVFoundation
+import Foundation
+import UIKit
+
+extension ARViewController: AVCaptureVideoDataOutputSampleBufferDelegate {
+    func setupAVCapture() {
+        session.sessionPreset = AVCaptureSessionPreset640x480
+        guard let device = AVCaptureDevice
+            .defaultDevice(withDeviceType: .builtInWideAngleCamera,
+                           mediaType: AVMediaTypeVideo,
+                           position: .back) else{
+                            return
+        }
+        captureDevice = device
+        beginSession()
+        done = true
+    }
+    
+    func beginSession() {
+        var deviceInput: AVCaptureDeviceInput?
+        do {
+            deviceInput = try AVCaptureDeviceInput(device: captureDevice)
+        } catch let error {
+            deviceInput = nil
+            print("error: \(error.localizedDescription)")
+        }
+        if self.session.canAddInput(deviceInput) {
+            self.session.addInput(deviceInput)
+        }
+        
+        videoDataOutput = AVCaptureVideoDataOutput()
+        videoDataOutput.alwaysDiscardsLateVideoFrames = true
+        videoDataOutputQueue = DispatchQueue(label: "VideoDataOutputQueue")
+        videoDataOutput.setSampleBufferDelegate(self, queue: videoDataOutputQueue)
+        if session.canAddOutput(videoDataOutput) {
+            session.addOutput(videoDataOutput)
+        }
+        videoDataOutput.connection(withMediaType: AVMediaTypeVideo).isEnabled = true
+        
+        cameraViewLayer = AVCaptureVideoPreviewLayer(session: session)
+        cameraViewLayer.videoGravity = AVLayerVideoGravityResizeAspect
+        
+        let rootLayer = cameraView.layer
+        rootLayer.masksToBounds = true
+        cameraViewLayer.frame = rootLayer.bounds
+        rootLayer.addSublayer(cameraViewLayer)
+        session.startRunning()
+    }
+    
+    func captureOutput(_ captureOutput: AVCaptureOutput!, didOutputSampleBuffer sampleBuffer: CMSampleBuffer!, from connection: AVCaptureConnection!) {
+        currentFrame = convertImageFromCMSampleBufferRef(sampleBuffer)
+    }
+    
+    // clean up AVCapture
+    func stopCamera(){
+        session.stopRunning()
+        done = false
+    }
+    
+    func convertImageFromCMSampleBufferRef(_ sampleBuffer:CMSampleBuffer) -> CIImage {
+        let pixelBuffer = CMSampleBufferGetImageBuffer(sampleBuffer)!
+        let ciImage = CIImage(cvImageBuffer: pixelBuffer)
+        return ciImage
+    }
+}
